MAINT, DEP: Restructure the `qmflows.packages` sub-package
Restructure the sub-package into the following public namespaces:
* `packages`

Also deprecates four functions:
* `registry`
* `load_properties`
* `SerMolecule`
* `SerSettings`
