+++ .github/workflows/pythonapp.yml
@@ -5,6 +5,11 @@ on:
     branches:
       - master
   pull_request:
+  workflow_dispatch:
+
+concurrency:
+    group: ${{ github.workflow }}-${{ github.head_ref || github.run_id }}
+    cancel-in-progress: true
 
 jobs:
   Test:
@@ -38,12 +43,7 @@ jobs:
       CP2K_DATA_DIR: "/home/runner/work/qmflows/qmflows/cp2k/data"
 
     steps:
-    - name: Cancel Previous Runs
-      uses: styfle/cancel-workflow-action@0.9.1
-      with:
-        access_token: ${{ github.token }}
-
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v3
 
     - name: Install cp2k
       if: matrix.os == 'ubuntu-latest'
@@ -53,13 +53,17 @@ jobs:
             VERSION_SHORT=$(echo "${{ matrix.special }}" | grep -P -o "CP2K \K([\.0-9]+)")
             VERSION=v"$VERSION_SHORT".0
 
+            echo ::group::"Get CP2K $VERSION data"
             git clone https://github.com/cp2k/cp2k -q
             cd cp2k
             git -c advice.detachedHead=false checkout tags/$VERSION
+            echo ::endgroup::
 
+            echo ::group::"Get CP2K $VERSION binaries"
             wget -nv https://github.com/cp2k/cp2k/releases/download/$VERSION/cp2k-$VERSION_SHORT-Linux-x86_64.ssmp -O cp2k.ssmp
             chmod u+rx cp2k.ssmp
             mv cp2k.ssmp /usr/local/bin/cp2k.popt
+            echo ::endgroup::
             ;;
           *)
             sudo apt install cp2k ;;
@@ -73,7 +77,6 @@ jobs:
       uses: s-weigand/setup-conda@v1
       with:
         update-conda: true
-        conda-channels: anaconda, conda-forge
 
     - name: Install dependencies
       shell: bash
@@ -84,7 +87,7 @@ jobs:
             source $CONDA/bin/activate test
             pip install rdkit-pypi
             pip install --pre -e .[test] --upgrade --force-reinstall
-            pip install git+https://github.com/SCM-NV/PLAMS@master --upgrade
+            pip uninstall plams -y; pip install git+https://github.com/SCM-NV/PLAMS@master --upgrade
             pip install git+https://github.com/NLeSC/noodles@master --upgrade
             ;;
           "no optional")
@@ -121,7 +124,7 @@ jobs:
             pytest -m 'not (slow or long)' ;;
         esac
 
-    - uses: codecov/codecov-action@v2
+    - uses: codecov/codecov-action@v3
       with:
         file: ./coverage.xml
         name: codecov-umbrella
@@ -129,15 +132,15 @@ jobs:
   Linting:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
 
       - name: Set up Python 3.10 on ubuntu-latest
-        uses: actions/setup-python@v2
+        uses: actions/setup-python@v3
         with:
           python-version: "3.10"
 
       - name: Install linters
-        run: pip install pydocstyle pycodestyle mypy "numpy>=1.21" pytest more_itertools types-PyYAML types-setuptools
+        run: pip install pydocstyle pycodestyle mypy "numpy>=1.21" pytest more_itertools types-PyYAML types-setuptools "pyparsing>=3.0.8"
 
       - name: Python info
         run: |
