+++ docs/_packages.rst
@@ -1,11 +1,11 @@
 qmflows.packages
 ================
 
-The :class:`~qmflows.packages.packages.Package` (sub-)classes of QMFlows.
+The :class:`~qmflows.packages.Package` (sub-)classes of QMFlows.
 
 Package-related Functions
 -------------------------
-.. currentmodule:: qmflows.packages.packages
+.. currentmodule:: qmflows.packages
 .. autosummary::
     run
 
@@ -14,7 +14,6 @@ The Package Class
 .. autosummary::
     Package
     Package.__init__
-    Package.__repr__
     Package.__call__
     Package.prerun
     Package.run_job
@@ -22,73 +21,104 @@ The Package Class
     Package.generic2specific
     Package.handle_special_keywords
 
-Package Subclasses
-------------------
-.. currentmodule:: qmflows.packages
 .. autosummary::
-    ~SCM.ADF
-    ~SCM.DFTB
-    ~cp2k_package.CP2K
-    ~cp2k_mm.CP2KMM
-    ~orca.ORCA
-    ~package_wrapper.PackageWrapper
+    ADF
+    DFTB
+    CP2K
+    CP2KMM
+    ORCA
+    PackageWrapper
 
 Package Instances
 ------------------
 .. autosummary::
-    ~SCM.adf
-    ~SCM.dftb
-    ~cp2k_package.cp2k
-    ~cp2k_mm.cp2k_mm
-    ~orca.orca
+    adf
+    dftb
+    cp2k
+    cp2k_mm
+    orca
+
+The Result Class
+-----------------
+.. autosummary::
+    Result
+    Result.__init__
+    Result.get_property
+    Result.results
+
+.. autosummary::
+    ADF_Result
+    DFTB_Result
+    CP2K_Result
+    CP2KMM_Result
+    ORCA_Result
+    ResultWrapper
 
 API
 ---
-.. autofunction:: qmflows.packages.packages.run
+.. autofunction:: qmflows.packages.run
 
 |
 
-.. autoclass:: qmflows.packages.packages.Package
+.. autoclass:: qmflows.packages.Package
     :members: generic_mapping, result_type, pkg_name
 
-.. automethod:: qmflows.packages.packages.Package.__init__
-.. automethod:: qmflows.packages.packages.Package.__repr__
-.. automethod:: qmflows.packages.packages.Package.__call__
-.. automethod:: qmflows.packages.packages.Package.prerun
-.. automethod:: qmflows.packages.packages.Package.run_job
-.. automethod:: qmflows.packages.packages.Package.postrun
-.. automethod:: qmflows.packages.packages.Package.generic2specific
-.. automethod:: qmflows.packages.packages.Package.handle_special_keywords
+.. automethod:: qmflows.packages.Package.__init__
+.. automethod:: qmflows.packages.Package.__call__
+.. automethod:: qmflows.packages.Package.prerun
+.. automethod:: qmflows.packages.Package.run_job
+.. automethod:: qmflows.packages.Package.postrun
+.. automethod:: qmflows.packages.Package.generic2specific
+.. automethod:: qmflows.packages.Package.handle_special_keywords
 
 |
 
-.. autoclass:: qmflows.packages.SCM.ADF
+.. autoclass:: qmflows.packages.ADF
 
 |
 
-.. autoclass:: qmflows.packages.SCM.DFTB
+.. autoclass:: qmflows.packages.DFTB
 
 |
 
-.. autoclass:: qmflows.packages.cp2k_package.CP2K
+.. autoclass:: qmflows.packages.CP2K
 
 |
 
-.. autoclass:: qmflows.packages.cp2k_mm.CP2KMM
+.. autoclass:: qmflows.packages.CP2KMM
 
 |
 
-.. autoclass:: qmflows.packages.orca.ORCA
+.. autoclass:: qmflows.packages.ORCA
 
 |
 
-.. autoclass:: qmflows.packages.package_wrapper.PackageWrapper
+.. autoclass:: qmflows.packages.PackageWrapper
     :noindex:
 
 |
 
-.. autodata:: qmflows.packages.SCM.adf
-.. autodata:: qmflows.packages.SCM.dftb
-.. autodata:: qmflows.packages.cp2k_package.cp2k
-.. autodata:: qmflows.packages.cp2k_mm.cp2k_mm
-.. autodata:: qmflows.packages.orca.orca
+.. autoclass:: qmflows.packages.Result
+    :members: prop_mapping
+
+.. automethod:: qmflows.packages.Result.__init__
+.. automethod:: qmflows.packages.Result.get_property
+.. autoproperty:: qmflows.packages.Result.results
+
+|
+
+.. autoclass:: qmflows.packages.ADF_Result
+.. autoclass:: qmflows.packages.DFTB_Result
+.. autoclass:: qmflows.packages.CP2K_Result
+.. autoclass:: qmflows.packages.CP2KMM_Result
+.. autoclass:: qmflows.packages.ORCA_Result
+.. autoclass:: qmflows.packages.ResultWrapper
+    :noindex:
+
+|
+
+.. autofunction:: qmflows.packages.adf
+.. autofunction:: qmflows.packages.dftb
+.. autofunction:: qmflows.packages.cp2k
+.. autofunction:: qmflows.packages.cp2k_mm
+.. autofunction:: qmflows.packages.orca
