+++ src/qmflows/data/dictionaries/propertiesCP2K.yaml
@@ -4,7 +4,7 @@ energy:
     kwargs:
         pattern: Total energy
         pos: 2
-    parser: generic_parsers
+    parser: null
 frequencies:
     file_pattern: '*-VIBRATIONS-1.mol'
     function: get_cp2k_freq
