+++ src/qmflows/data/dictionaries/propertiesORCA.yaml
@@ -3,31 +3,31 @@ charges:
     function: awk_file
     kwargs:
         script: /ATOM       NA         ZA/ {getline; while ($0 != "") {print $5;getline}}
-    parser: generic_parsers
+    parser: null
 dipole:
     file_ext: out
     function: awk_file
     kwargs:
         script: /Total Dipole Moment/ {print $5; print $6; print $7}
-    parser: generic_parsers
+    parser: null
 energy:
     file_ext: out
     function: awk_file
     kwargs:
         script: /FINAL SINGLE POINT ENERGY/ {energy = $5} END {print energy}
-    parser: generic_parsers
+    parser: null
 enthalpy:
     file_ext: out
     function: awk_file
     kwargs:
         script: /Total enthalpy/ {enthalpy = $4} END {print enthalpy}
-    parser: generic_parsers
+    parser: null
 free_energy:
     file_ext: out
     function: awk_file
     kwargs:
         script: /Final Gibbs free enthalpy/ {energy = $6} END {print energy}
-    parser: generic_parsers
+    parser: null
 frequencies:
     file_ext: hess
     function: parse_frequencies
@@ -47,7 +47,7 @@ optcycles:
     function: awk_file
     kwargs:
         script: /GEOMETRY OPTIMIZATION CYCLE/ {ncycles = $5} END {print ncycles}
-    parser: generic_parsers
+    parser: null
 orbitals:
     file_ext: out
     function: parse_molecular_orbitals
@@ -57,4 +57,4 @@ runtime:
     function: awk_file
     kwargs:
         script: /Sum of individual times/ {runtime = $6} END {print runtime}
-    parser: generic_parsers
+    parser: null
