+++ test/test_molecule.py
@@ -1,5 +1,6 @@
 """Test molecule serialization."""
-from qmflows import packages
+
+from qmflows.packages._packages import registry
 from qmflows.test_utils import PATH
 from scm.plams import Molecule
 
@@ -10,7 +11,7 @@ WATER.guess_bonds()
 def test_SerMolecule():
     """Test molecule serialization."""
     mol = WATER
-    registry = packages.registry()
-    encoded_molecule = registry.deep_encode(mol)
-    decoded_molecule = registry.deep_decode(encoded_molecule)
+    reg = registry()
+    encoded_molecule = reg.deep_encode(mol)
+    decoded_molecule = reg.deep_decode(encoded_molecule)
     assert len(mol) == len(decoded_molecule)
